requirements.txt
discord.py
requests~=2.24.0
python-dateutil~=2.8.1
